telegram-bot-chat/requirements.txt
python-telegram-bot
httpx
dotenv
openai
